requirements.txt
numpy
pandas
sklearn
numpy
pandas
seaborn
sklearn
pymongo
fastapi
notebook
notebook_shim
matplotlib
matplotlib-inline
ipykernel
PyYAML
dill
uvicorn
TTS
mlflow
tqdm
subprocess.run